Store player level in a short

+++ src/game/spell/Cheat.cpp
@@ -366,7 +366,7 @@ void ApplySPMax() {
 		ARX_SPEECH_Add(text);
 		player.Attribute_Redistribute+=10;
 		player.Skill_Redistribute+=50;
-		player.level=std::max((int)player.level,10);
+		player.level = std::max(short(player.level), short(10));
 		player.xp=GetXPforLevel(10);
 	}
 	else
